revert: move SignItem back to types
