+++ privval/dash_consensus_key.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/dashpay/dashd-go/btcjson"
 
-	"github.com/dashpay/tenderdash/crypto"
 	tmcrypto "github.com/dashpay/tenderdash/crypto"
+	"github.com/dashpay/tenderdash/types"
 )
 
 type dashConsensusPrivateKey struct {
@@ -105,7 +105,7 @@ func (pub DashConsensusPublicKey) VerifySignature(msg []byte, sig []byte) bool {
 	return pub.VerifySignatureDigest(hash, sig)
 }
 func (pub DashConsensusPublicKey) VerifySignatureDigest(hash []byte, sig []byte) bool {
-	signID := crypto.NewSignItemFromHash(pub.quorumType, pub.quorumHash, hash, hash).SignHash
+	signID := types.NewSignItemFromHash(pub.quorumType, pub.quorumHash, hash, hash).SignHash
 
 	return pub.PubKey.VerifySignatureDigest(signID, sig)
 }
